More debugging

+++ res/res_rtp_asterisk.c
@@ -1886,7 +1886,7 @@ static void process_dtmf_rfc2833(struct ast_rtp_instance *instance, unsigned cha
 				}
 #else
 				ast_debug(4, "--GOT DTMF END message. Duration samples %d (%ld ms)\n", rtp->dtmf_duration, f->len);
-				ast_verbose("-- Adress till option_dtmfminduration = %x option_debug = %x\n", &option_dtmfminduration, &option_debug);
+				ast_verbose("-- Adress till option_dtmfminduration = %x option_debug = %x\n", (int) &option_dtmfminduration, (int) &option_debug);
 #endif
 				rtp->resp = 0;
 				rtp->dtmf_duration = rtp->dtmf_timeout = 0;
